Deprecate WinRT helpers
