+++ lib/win32.dart
@@ -31,6 +31,7 @@ export 'src/callbacks.dart';
 export 'src/constants.dart';
 export 'src/constants_metadata.dart';
 export 'src/constants_nodoc.dart';
+export 'src/enums.dart';
 export 'src/exceptions.dart';
 export 'src/guid.dart';
 export 'src/inline.dart';
