+++ test/winrt_helpers_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 @TestOn('windows')
 
 import 'dart:ffi';
